Moved ProjectItem into it's own file
